BREAK AsioCallbackQueue::replaceGlobalQueue now takes a std::shared_ptr<asio::io_context>
The AsioCallbackQueue is a singleton so it is initialized as static and
detroyed after the end of main which used to own the io_context.
Using a shared ptr allows the AsioCallbackQueue to take the ownership
of the io_context after the end of main and avoid a crash when
Ctrl+C a node.